refactor test scenarios
