+++ src/tests/scenario/mod.rs
@@ -4,8 +4,9 @@ use testresult::TestResult;
 
 use crate::tests::network::EthereumConfig;
 
+pub mod beacon;
 pub mod erc20;
-pub mod finality;
+pub mod relayer;
 
 pub trait Scenario {
     fn run(&self, config: EthereumConfig) -> impl Future<Output = TestResult> + Send;
